+++ src/sage/symbolic/ginac/mul.h
@@ -53,8 +53,8 @@ public:
 	unsigned precedence() const override {return 50;}
 	bool info(unsigned inf) const override;
 	bool is_polynomial(const ex & var) const override;
-	int degree(const ex & s) const override;
-	int ldegree(const ex & s) const override;
+	numeric degree(const ex & s) const override;
+	numeric ldegree(const ex & s) const override;
 	ex coeff(const ex & s, const ex & n) const override;
 	bool has(const ex & other, unsigned options = 0) const override;
 	ex eval(int level=0) const override;
